Fix double coordinate transformation in QuickDraw rendering
The coordinate transformation from LOCAL to GLOBAL was happening twice:
1. DrawPrimitive (QuickDrawCore.c) added portBits.bounds offset
2. QDPlatform_DrawShape (QuickDrawPlatform.c) added it again

This caused all window content to be drawn offset by the window's
position, appearing "down and to the right" of where it should be.

Fixed by removing the offset addition in QDPlatform_DrawShape since
the rect parameter is already in GLOBAL coordinates when it arrives.