Fix startup sound to play chime instead of beep
The PlayStartupSound() function was incorrectly playing kSoundEffectBeep
(1000Hz beep) instead of kSoundEffectStartupChime during the welcome screen.

This was the source of the 'beep' that was happening on boot. Now it will
play the classic System 7 startup chime sound, and since we disabled the
fallback beep for the chime, you should only hear the chime sound.

This fixes the issue where users heard 'beep and then chime'.